% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspme-package.R
\docType{package}
\name{bspme-package}
\alias{bspme}
\alias{bspme-package}
\title{bspme: Bayesian Spatial Measurement Error Models}
\description{
Scalable methods for fitting Bayesian linear and generalized linear models in the presence of spatial exposure measurement error. These models typically arise from a two-stage Bayesian analysis of environmental exposures and health outcomes. From a first-stage model, predictions of the covariate of interest (''exposure'') and their uncertainty information (typically contained in MCMC samples) are used to form a multivariate normal prior distribution for exposure in a second-stage regression model. This package also provides implementation of the methods used in Lee et al. (2024) \url{https://arxiv.org/abs/2401.00634}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://changwoo-lee.github.io/bspme/}
  \item Report bugs at \url{https://github.com/changwoo-lee/bspme/issues}
}

}
\author{
\strong{Maintainer}: Changwoo Lee \email{c.lee@stat.tamu.edu}

Authors:
\itemize{
  \item Eun Sug Park
}

}
\keyword{internal}
