% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspme-package.R
\name{health_sim}
\alias{health_sim}
\title{Simulated health data}
\format{
A data frame with n = 2000 rows and 6 variables:
\describe{
\item{Y}{simulated continuous health outcome}
\item{Ybinary}{simulated binary health outcome}
\item{lon}{simulated health subject longitude}
\item{lat}{simulated health subject latitude}
\item{Z}{simulated covariate (p=1) that is not subject to measurement error}
\item{X_true}{true ln(NO2) exposure used for simulating health outcome}
}
}
\usage{
data(health_sim)
}
\description{
Simulated health data associated with ln(NO2) concentration on Jan 10, 2012. For details, see \code{health_sim.R}.
}
