% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspme-package.R
\name{NO2_Jan2012}
\alias{NO2_Jan2012}
\title{Daily average NO2 concentrations in and around the Harris County, Texas, in Jan 2012}
\format{
A data frame with 651 (21 sites x 31 days) rows and 5 variables:
\describe{
\item{date}{date in POSIXct format}
\item{site_name}{monitoring station name}
\item{lon}{monitoring station longitude}
\item{lat}{monitoring station latitude}
\item{lnNO2}{natural logarithm of daily average NO2 concentrations measured in parts per billion by volume (ppbv)}
}
}
\usage{
data(NO2_Jan2012)
}
\description{
This dataset contains daily average NO2 (nitrogen dioxide) concentrations obtained from 21 monitoring stations in and around Harris County, Texas, in January 2012.
}
